# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Believer Backend API
  version: 1.8.0
  description: Active REST API contract for auth, account recovery, account-governance settings actions, business listings, mosques, backend-owned mosque prayer times, mosque page content, mosque image uploads, bookmarks, notifications, and services.
servers:
  - url: http://localhost:4000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request payload
        details:
          type: array
          items:
            type: object
    ErrorResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          type: object
          example: {}
...
    BusinessListingStatus:
      type: string
      enum: [draft, under_review, live, rejected]
    BusinessListingSubmitter:
      type: object
      required: [id, fullName, email]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
          format: email
    TimeOfDay:
      type: object
      required: [hour, minute]
      properties:
        hour:
          type: integer
          minimum: 0
          maximum: 23
        minute:
          type: integer
          minimum: 0
          maximum: 59
    BusinessListingLogo:
      type: object
      properties:
        fileName:
          type: string
          nullable: true
        bytesBase64:
          type: string
          nullable: true
        contentType:
          type: string
          nullable: true
        tileBackgroundColor:
          type: integer
          example: 4293512350
    BusinessListingSelectedType:
      type: object
      required: [groupId, groupLabel, itemId, itemLabel]
      properties:
        groupId:
          type: string
        groupLabel:
          type: string
        itemId:
          type: string
        itemLabel:
          type: string
    BusinessListingBasicDetails:
      type: object
      required: [businessName, logo, selectedType, tagline, description]
      properties:
        businessName:
          type: string
          maxLength: 180
        logo:
          allOf:
            - $ref: '#/components/schemas/BusinessListingLogo'
          nullable: true
        selectedType:
          allOf:
            - $ref: '#/components/schemas/BusinessListingSelectedType'
          nullable: true
        tagline:
          type: string
          maxLength: 160
        description:
          type: string
          maxLength: 4000
    BusinessListingBasicDraftDetails:
      type: object
      properties:
        businessName:
          type: string
          maxLength: 180
        logo:
          allOf:
            - $ref: '#/components/schemas/BusinessListingLogo'
          nullable: true
        selectedType:
          allOf:
            - $ref: '#/components/schemas/BusinessListingSelectedType'
          nullable: true
        tagline:
          type: string
          maxLength: 160
        description:
          type: string
          maxLength: 4000
    BusinessListingContactDetails:
      type: object
      required:
        [
          businessEmail,
          phone,
          whatsapp,
          openingTime,
          closingTime,
          instagramUrl,
          facebookUrl,
          websiteUrl,
          address,
          zipCode,
          city,
          onlineOnly
        ]
      properties:
        businessEmail:
          type: string
          maxLength: 254
        phone:
          type: string
          maxLength: 40
        whatsapp:
          type: string
          maxLength: 40
        openingTime:
          allOf:
            - $ref: '#/components/schemas/TimeOfDay'
          nullable: true
        closingTime:
          allOf:
            - $ref: '#/components/schemas/TimeOfDay'
          nullable: true
        instagramUrl:
          type: string
          maxLength: 240
        facebookUrl:
          type: string
          maxLength: 240
        websiteUrl:
          type: string
          maxLength: 240
        address:
          type: string
          maxLength: 240
        zipCode:
          type: string
          maxLength: 20
        city:
          type: string
          maxLength: 120
        onlineOnly:
          type: boolean
    BusinessListingContactDraftDetails:
      type: object
      properties:
        businessEmail:
          type: string
          maxLength: 254
        phone:
          type: string
          maxLength: 40
        whatsapp:
          type: string
          maxLength: 40
        openingTime:
          allOf:
            - $ref: '#/components/schemas/TimeOfDay'
          nullable: true
        closingTime:
          allOf:
            - $ref: '#/components/schemas/TimeOfDay'
          nullable: true
        instagramUrl:
          type: string
          maxLength: 240
        facebookUrl:
          type: string
          maxLength: 240
        websiteUrl:
          type: string
          maxLength: 240
        address:
          type: string
          maxLength: 240
        zipCode:
          type: string
          maxLength: 20
        city:
          type: string
          maxLength: 120
        onlineOnly:
          type: boolean
    BusinessListingDraftRequest:
      type: object
      required: [basicDetails, contactDetails]
      properties:
        basicDetails:
          $ref: '#/components/schemas/BusinessListingBasicDraftDetails'
        contactDetails:
          $ref: '#/components/schemas/BusinessListingContactDraftDetails'
    BusinessListingSubmissionRequest:
      type: object
      required: [basicDetails, contactDetails]
      properties:
        basicDetails:
          $ref: '#/components/schemas/BusinessListingBasicDetails'
        contactDetails:
          $ref: '#/components/schemas/BusinessListingContactDetails'
    BusinessListing:
      type: object
      required:
        [
          id,
          status,
          basicDetails,
          contactDetails,
          submittedAt,
          publishedAt,
          reviewedBy,
          reviewedAt,
          rejectionReason,
          createdAt,
          lastUpdatedAt
        ]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/BusinessListingStatus'
        basicDetails:
          $ref: '#/components/schemas/BusinessListingBasicDetails'
        contactDetails:
          $ref: '#/components/schemas/BusinessListingContactDetails'
        submittedAt:
          type: string
          format: date-time
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        reviewedBy:
          type: string
          format: uuid
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        submitter:
          allOf:
            - $ref: '#/components/schemas/BusinessListingSubmitter'
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
        lastUpdatedAt:
          type: string
          format: date-time
          nullable: true
    BusinessListingPayload:
      type: object
      required: [listing]
      properties:
        listing:
          allOf:
            - $ref: '#/components/schemas/BusinessListing'
          nullable: true
    BusinessListingEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: '#/components/schemas/BusinessListingPayload'
        error:
          nullable: true
          example: null
        meta:
          type: object
          example: {}
    BusinessListingListPayload:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BusinessListing'
    BusinessListingListEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: '#/components/schemas/BusinessListingListPayload'
        error:
          nullable: true
          example: null
        meta:
          type: object
          example: {}
    BusinessListingRejectionRequest:
      type: object
      required: [rejectionReason]
      properties:
        rejectionReason:
          type: string
          minLength: 3
          maxLength: 1000
...
paths:
...
  /api/v1/business-listings/draft:
    put:
      tags: [Business Listings]
      summary: Create or update the current user's business listing draft
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessListingDraftRequest'
      responses:
        '200':
          description: Draft updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessListingEnvelope'
        '201':
          description: Draft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessListingEnvelope'
        '400':
          description: Invalid draft payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/business-listings/submit:
    post:
      tags: [Business Listings]
      summary: Submit the current user's business listing for review
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessListingSubmissionRequest'
      responses:
        '202':
          description: Listing submitted for review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessListingEnvelope'
        '400':
          description: Invalid submission payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/business-listings/me:
    get:
      tags: [Business Listings]
      summary: Read the current user's latest business listing status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Latest listing status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessListingEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/business-listings/pending:
    get:
      tags: [Business Listings]
      summary: List business listings waiting for super-admin moderation
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pending business listings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessListingListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Super-admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/business-listings/{id}:
    get:
      tags: [Business Listings]
      summary: Read one business listing for super-admin moderation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Business listing detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessListingEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Super-admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/business-listings/{id}/approve:
    post:
      tags: [Business Listings]
      summary: Approve a business listing as a super admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Business listing approved and published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessListingEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Super-admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Listing is no longer under review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/business-listings/{id}/reject:
    post:
      tags: [Business Listings]
      summary: Reject a business listing as a super admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessListingRejectionRequest'
      responses:
        '200':
          description: Business listing rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessListingEnvelope'
        '400':
          description: Invalid rejection payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Super-admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Listing is no longer under review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'